+++ changelog.md
@@ -11,7 +11,10 @@ latest
  - Prevent slicing by transient assigning an ot meta loop point flag.
  - Added button to set a custom slice count on the slices panel.
  - Change default chain names to not contain _ or [ ] chars.
- - Turn off embed slice data by default for new instances of web app.
+ - Turn off deprecated embed slice data config option by default.
+ - Use Cue markers as preferred slice data storage within the wav file data.
+ - Turn on/off .ot file export options when choosing the ot common configuration on the audio config panel.
+ - Add OP-XY to common configurations, as this uses wav files as the primary file type.
 
 2024.05.28 1.4.15
  - Amplitude threshold condense selected action, enter an upper and lower amplitude value (between 0 and 1), the sample will be modified to only include amplitudes within the given range, gaps between the audio is removed condensing the sample size.